requirements.txt
discord==0.0.2
discord.py==0.16.12
peewee>=3.8.0
python-dotenv==0.8.2
